debug_bundle: Generate and store metadata file
The purpose of the metadata file is to provide information about the run
that generated the debug bundle file that is present in the same
directory as the metadata file.  It will be used to build out the status
return from the service in the event that the service is restarted.  It
also contains the SHA256 checksum to validate that the debug bundle has
not been modified.

Signed-off-by: Michael Boquard <michael@redpanda.com>