cl/model: Added add and update mirror topic cmds
Added two new commands to handle creation of mirror topics and updating
the mirror topic state.  Though they are currently the same values, the
creation command may be updated to also add in fields such as selecting
the name of the destination topic if it differs from the source topic
name.

Signed-off-by: Michael Boquard <michael@redpanda.com>