cl: Update source_topic_syncer to sync topic properties
This updates the source_topic_syncer task to also update the properties
of existing mirror topics based on the properties of the source topic.

This runs before attempting to create new topics.  This will step
through existing mirror topics and check the source topic's metadata and
properties and, if they have changed, update the mirror topic in the
cluster link table.  It will mark mirror topics as failed if the
partition count has gone backwards.  Once Topic IDs are supported in
Redpanda, we will also mark those topics as failed if their topic IDs
have changed.

Signed-off-by: Michael Boquard <michael@redpanda.com>