sr: Check SR is not being mirrored before writing
Create a new state checker interface class that is provided to
seq_writer that will return whether or not the schemas topic is
writable.  If not, then the seq_writer will throw an exception.

This is wired in at the API to check to see if the schemas topic is
being mirrored by the Schema Registry. If it is, writes will be
disabled.

Signed-off-by: Michael Boquard <michael@redpanda.com>