cl: Check if _schemas is empty before mirroring
If the mirror_schema_registry_topic option is true, then the
source_topic_syncer task will attempt to add the schemas topic as a
mirror topic.

It is possible for the _schemas topic to already be present on the
Shadow Cluster before Shadow Linking starts.  The SR service will
attempt to create the topic upon the first access of its API endpoint.
The task will check to see if the topic exists, and if it does, if it's
empty.  If it doesn't exist, or if it is empty, then we will add the
schemas topic to the mirror topic list within the cluster link table.

Future commits will prevent the schema registry service from writing
data to the topic if that configuration option is true or if the
_schemas topic is in the mirror topic list.

Signed-off-by: Michael Boquard <michael@redpanda.com>