cl: Call start replicators on newly added shadow topics
If shadow topics are added to the link, check if they are active and if
they have local leadership.  If so, attempt to start replication on
them.  This is needed for existing topics that become shadow topics.
The changes previously in replicator ensure that starts from link to not
interfere with starts due to partition leadership changes.

Signed-off-by: Michael Boquard <michael@redpanda.com>